requirements.txt
absl-py==0.8.1
astor==0.8.1
atari-py==0.2.6
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
backcall==0.1.0
baselines==0.1.5
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
Cython==0.29.14
dateparser==0.7.2
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
entrypoints==0.3
future==0.18.2
gast==0.2.2
glfw==1.8.7
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
gym==0.15.4
h5py==2.10.0
hyperlink==19.0.0
idna==2.8
imageio==2.6.1
importlib-metadata==1.2.0
incremental==17.5.0
ipykernel==5.1.3
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.6.1
Keras==2.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lockfile==0.12.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mock==3.0.5
more-itertools==7.2.0
mpi4py==3.0.3
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
progressbar2==3.47.0
prometheus-client==0.7.1
protobuf==3.11.1
psutil==5.6.7
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pyglet==1.3.2
Pygments==2.5.2
PyHamcrest==1.9.0
pymongo==3.10.0
pyOpenSSL==19.1.0
pyparsing==2.4.5
pyrsistent==0.15.6
python-binance==0.7.4
python-dateutil==2.8.1
python-utils==2.3.0
pytz==2019.3
PyYAML==5.2
pyzmq==18.1.0
qtconsole==4.6.0
regex==2019.12.19
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22
scipy==1.3.3
Send2Trash==1.5.0
service-identity==18.1.0
six==1.13.0
stable-baselines==2.8.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.40.2
traitlets==4.3.3
Twisted==19.10.0
txaio==18.8.1
tzlocal==2.0.0
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
zmq==0.0.0
zope.interface==4.7.1
